// src/ui/hud/targeting/types.ts
export interface LockCard {
  el: HTMLElement;
  headerEl: HTMLElement;
  iconEl: HTMLCanvasElement;
  canvasEl: HTMLCanvasElement;
  nameEl: HTMLElement;
  levelEl: HTMLElement;
  targetIndEl: HTMLElement;
  barsEl: HTMLElement;
  shieldInner: HTMLElement;
  shieldLabel: HTMLElement;
  hpInner: HTMLElement;
  hpLabel: HTMLElement;
  structInner: HTMLElement;
  structLabel: HTMLElement;
  telemetryEl: HTMLElement;
  spdMetric: HTMLElement;
  distMetric: HTMLElement;
  sigMetric: HTMLElement;
  trsMetric: HTMLElement;
  metaEl: HTMLElement;
  scanEl: HTMLElement;
  assignEl: HTMLElement;
}